% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_owner_repo.R
\name{add_owner_repo}
\alias{add_owner_repo}
\title{Add owner/repo}
\usage{
add_owner_repo(
  dt,
  add_ref = TRUE,
  ref_cols = c("ref", "owner_repo", "repo", "package", "Package"),
  sep = "/"
)
}
\arguments{
\item{dt}{\link[data.table]{data.table}.}

\item{add_ref}{Coalesce multiple columns 
(\code{e.g. c("owner_repo","repo","package","Package")}) into a single "ref"
column. Tries to fill in \code{NA} values as much as possible.}

\item{ref_cols}{Columns to consider when coalescing into the "ref" column.}

\item{sep}{Separator value between "owner" and "repo".}
}
\value{
\link[data.table]{data.table}.
}
\description{
Try to add a new column named "owner_repo" that combines both the owner of 
and name of each GitHub repository.
}
\examples{
dt <- data.table::data.table(owner=letters, repo=LETTERS)
dt <- add_owner_repo(dt)
}
