% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description_authors.R
\name{description_authors}
\alias{description_authors}
\title{Description authors}
\usage{
description_authors(
  desc_file,
  names_only = TRUE,
  add_html = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{desc_file}{When \code{owner} or \code{repo} are NULL, 
these arguments are inferred from the \emph{DESCRIPTION} file.}

\item{names_only}{Only return author names (and not other subfields).
For example:
"Jim Hester \<james.f.hester\@gmail.com\> \[aut\]" would become
"Jim Hester".}

\item{add_html}{Add HTML styling to certain fields (e.g "authors").}

\item{verbose}{Print messages.}
}
\value{
Authors as an HTML character string.
}
\description{
Parse authors from the \emph{DESCRIPTION} file of an R package.
}
\keyword{internal}
