% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description_extract.R
\name{description_extract}
\alias{description_extract}
\title{DESCRIPTION file: extract fields}
\usage{
description_extract(
  desc_file = NULL,
  ref = NULL,
  fields = c("owner", "repo", "authors"),
  names_only = TRUE,
  add_html = FALSE,
  as_datatable = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{desc_file}{When \code{owner} or \code{repo} are NULL, 
these arguments are inferred from the \emph{DESCRIPTION} file.}

\item{ref}{Reference for a GitHub repository. If \code{NULL}
(the default), the reference is determined by the URL
field in the DESCRIPTION file.}

\item{fields}{Fields to extract.}

\item{names_only}{Only return author names (and not other subfields).
For example:
"Jim Hester \<james.f.hester\@gmail.com\> \[aut\]" would become
"Jim Hester".}

\item{add_html}{Add HTML styling to certain fields (e.g "authors").}

\item{as_datatable}{Convert the results into a \link[data.table]{data.table}.}

\item{verbose}{Print messages.}
}
\value{
A named list or \link[data.table]{data.table}.
}
\description{
Extract information from an R package's DESCRIPTION file.
}
\examples{
 
res <- description_extract(ref="RajLabMSSM/echolocatoR")
}
