% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description_read.R
\name{description_read}
\alias{description_read}
\title{Read a \emph{DESCRIPTION} file}
\usage{
description_read(dcf, fields = NULL, drop = TRUE, encoding = "")
}
\arguments{
\item{dcf}{A path to a \emph{DESCRIPTION} file or a matrix object from 
\link[base]{read.dcf}.

Code adapted from the function \link[utils]{packageDescription}.}

\item{fields}{a character vector giving the tags of fields to return
    (if other fields occur in the file they are ignored).}

\item{drop}{If \code{TRUE} and the length of \code{fields} is 1, then
    a single character string with the value of the respective field is
    returned instead of an object of class \code{"packageDescription"}.}

\item{encoding}{If there is an \code{Encoding} field, to what encoding
    should re-encoding be attempted?  If \code{NA}, no re-encoding.  The
    other values are as used by \code{\link{iconv}}, so the default
    \code{""} indicates the encoding of the current locale.}
}
\description{
Read in a \emph{DESCRIPTION} file as a nicely organized
 \link[utils]{packageDescription} object.
}
\keyword{internal}
