% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_branches.R
\name{github_branches}
\alias{github_branches}
\title{GitHub branches}
\usage{
github_branches(
  owner = NULL,
  repo = NULL,
  branch = NULL,
  master_or_main = TRUE,
  as_datatable = FALSE,
  token = gh::gh_token(),
  desc_file = NULL,
  error = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{owner}{Owner of the GitHub repository. 
If \code{NULL}, will automatically try to infer the owner
 name from the \emph{DESCRIPTION file}
 (assuming you're working directory is a local R package repo).}

\item{repo}{GitHub repository name. 
If \code{NULL}, will automatically try to infer the repo name
 name from the \emph{DESCRIPTION file}
 (assuming you're working directory is a local R package repo).}

\item{branch}{[Optional] If \code{branch} is supplied 
(as a character vector of one or more branch names), 
will check to see if that branch exists. If it does, only that branch will
be returned. If it doesn't, an error will be thrown.}

\item{master_or_main}{If \code{branch} is supplied and
 is either \code{"master"} or \code{"main"},
automatically interpret "master" and "main" as synonymous and return 
whichever branch exists.}

\item{as_datatable}{Return the results as a \link[data.table]{data.table}
(\code{TRUE}), or as a character vector of branch names
 (default: \code{FALSE}).}

\item{token}{GitHub Personal Authentication Token (PAT).
See \href{https://docs.github.com/en/authentication/keeping-your-account-and-data-secure/creating-a-personal-access-token}{
here for further instructions}.}

\item{desc_file}{When \code{owner} or \code{repo} are NULL, 
these arguments are inferred from the \emph{DESCRIPTION} file.
If a  \emph{DESCRIPTION} file cannot be found in the current working directory, 
it then searches for one in the remote GitHub repository.}

\item{error}{Throw an error when no matching branches are fond.}

\item{verbose}{Print messages.}
}
\value{
Character vector or \link[data.table]{data.table} of branches.
}
\description{
List all branches for a given GitHub repository.
}
\examples{
branches <- github_branches(owner="RajLabMSSM", repo="echolocatoR")
}
