% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_code.R
\name{github_code}
\alias{github_code}
\title{GitHub code}
\source{
\href{https://docs.github.com/en/search-github/searching-on-github/searching-code}{
GitHub Docs: Searching code}
}
\usage{
github_code(query, token = gh::gh_token(), .limit = Inf, verbose = TRUE)
}
\arguments{
\item{query}{Regex query to filter files by.}

\item{token}{GitHub Personal Authentication Token (PAT).
See \href{https://docs.github.com/en/authentication/keeping-your-account-and-data-secure/creating-a-personal-access-token}{
here for further instructions}.}

\item{.limit}{Number of records to return. This can be used
instead of manual pagination. By default it is \code{NULL},
which means that the defaults of the GitHub API are used.
You can set it to a number to request more (or less)
records, and also to \code{Inf} to request all records.
Note, that if you request many records, then multiple GitHub
API calls are used to get them, and this can take a potentially
long time.}

\item{verbose}{Print messages.}
}
\description{
Search for GitHub code using specific queries.
}
\examples{
\dontrun{
## easily exceeds API limit
repos <- github_code(query="Package path:DESCRIPTION", .limit=5)
}
}
