% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_files_download.R
\name{github_files_download}
\alias{github_files_download}
\title{Download files from GitHub}
\usage{
github_files_download(
  filelist,
  download_dir = tempdir(),
  overwrite = FALSE,
  timeout = 5 * 60,
  nThread = 1,
  verbose = TRUE
)
}
\description{
Download files from GitHub, e.g. those that have been found with
\link[echodata]{github_files}.
}
\keyword{internal}
