% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_files_gh.R
\name{github_files_gh}
\alias{github_files_gh}
\title{List files in GitHub repo: via \pkg{gh}}
\source{
\href{https://docs.github.com/en/rest/overview/endpoints-available-for-github-apps}{
GitHub endpoints}

\href{https://docs.github.com/en/rest/git/trees#get-a-tree-recursively}{
GitHub trees API}
}
\usage{
github_files_gh(
  owner,
  repo,
  branch = c("master", "main"),
  token = gh::gh_token(),
  verbose = TRUE
)
}
\arguments{
\item{owner}{Repo owner name
(can be organization or individual GitHub account name).}

\item{repo}{GitHub repository name.}

\item{branch}{Which branch to search.}

\item{token}{GitHub Personal Authentication Token (PAT).
See \href{https://docs.github.com/en/authentication/keeping-your-account-and-data-secure/creating-a-personal-access-token}{
here for further instructions}.}

\item{verbose}{Print messages.}
}
\value{
A \link[data.table]{data.table} fo file paths.
}
\description{
Search for files within a public GitHub repository and return their paths.
}
\keyword{internal}
