% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_files_httr.R
\name{github_files_httr}
\alias{github_files_httr}
\title{List files in GitHub repo: via \pkg{httr}}
\usage{
github_files_httr(
  owner,
  repo,
  branch = c("main", "master"),
  seconds = 5 * 60,
  verbose = TRUE
)
}
\arguments{
\item{owner}{Repo owner name
(can be organization or individual GitHub account name).}

\item{repo}{GitHub repository name.}

\item{branch}{Which branch to search.}

\item{seconds}{number of seconds to wait for a response until giving up.
Can not be less than 1 ms.}

\item{verbose}{Print messages.}
}
\value{
A \link[data.table]{data.table} fo file paths.
}
\description{
Search for files within a public GitHub repository and return their paths.
}
\keyword{internal}
