% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_hex.R
\name{github_hex}
\alias{github_hex}
\title{GitHub hex}
\usage{
github_hex(
  owner = NULL,
  repo = NULL,
  branch = "master",
  master_or_main = TRUE,
  hex_path = "inst/hex/hex.png",
  desc_file = NULL,
  add_html = TRUE,
  height = 300,
  token = gh::gh_token(),
  verbose = TRUE
)
}
\arguments{
\item{owner}{Owner of the GitHub repository. 
If \code{NULL}, will automatically try to infer the owner
 name from the \emph{DESCRIPTION file}
 (assuming you're working directory is a local R package repo).}

\item{repo}{GitHub repository name. 
If \code{NULL}, will automatically try to infer the repo name
 name from the \emph{DESCRIPTION file}
 (assuming you're working directory is a local R package repo).}

\item{branch}{[Optional] If \code{branch} is supplied 
(as a character vector of one or more branch names), 
will check to see if that branch exists. If it does, only that branch will
be returned. If it doesn't, an error will be thrown.}

\item{master_or_main}{If \code{branch} is supplied and
 is either \code{"master"} or \code{"main"},
automatically interpret "master" and "main" as synonymous and return 
whichever branch exists.}

\item{hex_path}{Path to the hex sticker in the GitHub repo.}

\item{desc_file}{When \code{owner} or \code{repo} are NULL, 
these arguments are inferred from the \emph{DESCRIPTION} file.
If a  \emph{DESCRIPTION} file cannot be found in the current working directory, 
it then searches for one in the remote GitHub repository.}

\item{add_html}{Wrap the URL in HTML syntax (i.e. "<img src=...").}

\item{height}{Height of the HTML image (only used when \code{add_html=TRUE}).}

\item{token}{GitHub Personal Authentication Token (PAT).
See \href{https://docs.github.com/en/authentication/keeping-your-account-and-data-secure/creating-a-personal-access-token}{
here for further instructions}.}

\item{verbose}{Print messages.}
}
\value{
URL to hex sticker.
}
\description{
Find the URL to a hex sticker for an R package in a GitHub repository.
If none can be found, returns \code{NULL}.
}
\examples{
hex <- github_hex(owner="RajLabMSSM", repo="echolocatoR")
}
