% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_metadata.R
\name{github_metadata}
\alias{github_metadata}
\title{GitHub metadata}
\usage{
github_metadata(
  owner,
  repo,
  add_traffic = FALSE,
  token = gh::gh_token(),
  .limit = Inf,
  verbose = TRUE
)
}
\arguments{
\item{owner}{Repo owner name
(can be organization or individual GitHub account name).}

\item{repo}{GitHub repository name.}

\item{add_traffic}{Add traffic metadata 
with \link[echogithub]{github_traffic}.}

\item{token}{GitHub Personal Authentication Token (PAT).
See \href{https://docs.github.com/en/authentication/keeping-your-account-and-data-secure/creating-a-personal-access-token}{
here for further instructions}.}

\item{.limit}{Number of records to return. This can be used
instead of manual pagination. By default it is \code{NULL},
which means that the defaults of the GitHub API are used.
You can set it to a number to request more (or less)
records, and also to \code{Inf} to request all records.
Note, that if you request many records, then multiple GitHub
API calls are used to get them, and this can take a potentially
long time.}

\item{verbose}{Print messages.}
}
\value{
A \link[data.table]{data.table}.
}
\description{
Get metadata on a GitHub repository.
}
\examples{
\dontrun{
dt <- github_metadata(owner="RajLabMSSM",
                      repo="echolocatoR")
}
}
