% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_pages.R
\name{github_pages}
\alias{github_pages}
\title{GitHub Pages website}
\usage{
github_pages(owner, repo, error = TRUE, token = gh::gh_token(), verbose = TRUE)
}
\arguments{
\item{owner}{Repo owner name
(can be organization or individual GitHub account name).}

\item{repo}{GitHub repository name.}

\item{error}{Throw an error when the GitHub repository does not exist 
(default: \code{TRUE}).}

\item{token}{GitHub Personal Authentication Token (PAT).
See \href{https://docs.github.com/en/authentication/keeping-your-account-and-data-secure/creating-a-personal-access-token}{
here for further instructions}.}

\item{verbose}{Print messages.}
}
\description{
Return the \href{https://pages.github.com/}{GitHub Pages} website URL 
for a given repository, if one exists.
}
\examples{
link <- github_pages(owner="RajLabMSSM", repo="echolocatoR")
}
