% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_pages_files.R
\name{github_pages_files}
\alias{github_pages_files}
\title{Find pages}
\usage{
github_pages_files(
  owner,
  repo,
  branch = "gh-pages",
  subdir = NULL,
  query = NULL,
  local_repo = NULL,
  save_path = NULL,
  token = gh::gh_token(),
  .limit = Inf,
  verbose = TRUE
)
}
\arguments{
\item{owner}{Repo owner name
(can be organization or individual GitHub account name).}

\item{repo}{GitHub repository name.}

\item{branch}{Which branch to search.}

\item{subdir}{Only search for files within a given subdirectory 
(e.g. "docs").}

\item{query}{Regex query to filter files by.}

\item{local_repo}{Search a cloned local folder of a git repo instead of 
the remote GitHub repo. If not \code{NULL}, expects a path to the local repo.}

\item{save_path}{Path to save the results to (as a ".csv").}

\item{token}{GitHub Personal Authentication Token (PAT).
See \href{https://docs.github.com/en/authentication/keeping-your-account-and-data-secure/creating-a-personal-access-token}{
here for further instructions}.}

\item{.limit}{Number of records to return. This can be used
instead of manual pagination. By default it is \code{NULL},
which means that the defaults of the GitHub API are used.
You can set it to a number to request more (or less)
records, and also to \code{Inf} to request all records.
Note, that if you request many records, then multiple GitHub
API calls are used to get them, and this can take a potentially
long time.}

\item{verbose}{Print messages.}
}
\description{
Find html files on GitHub Pages.
}
\examples{
dt <- github_pages_files(owner="RajLabMSSM",
                         repo="echolocatoR", 
                         .limit=5)
}
