% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_pages_vignettes.R
\name{github_pages_vignettes}
\alias{github_pages_vignettes}
\title{GitHub Pages vignettes}
\usage{
github_pages_vignettes(
  owner,
  repo,
  as_toc = FALSE,
  branch = "gh-pages",
  subdir = "articles",
  query = "*.*.html$",
  local_repo = NULL,
  save_path = NULL,
  token = gh::gh_token(),
  verbose = FALSE
)
}
\arguments{
\item{owner}{Repo owner name
(can be organization or individual GitHub account name).}

\item{repo}{GitHub repository name.}

\item{as_toc}{Return the list of vignettes as a
 table of contents (TOC) HTML string.
\emph{NOTE}: If you want to show the TOC in an Rmarkdown file, 
make sure you change the code chunk settings to: 
\code{{r, results='asis'}}.}

\item{branch}{Which branch to search.}

\item{subdir}{Only search for files within a given subdirectory 
(e.g. "docs").}

\item{query}{Regex query to filter files by.}

\item{local_repo}{Search a cloned local folder of a git repo instead of 
the remote GitHub repo. If not \code{NULL}, expects a path to the local repo.}

\item{save_path}{Path to save the results to (as a ".csv").}

\item{token}{GitHub Personal Authentication Token (PAT).
See \href{https://docs.github.com/en/authentication/keeping-your-account-and-data-secure/creating-a-personal-access-token}{
here for further instructions}.}

\item{verbose}{Print messages.}
}
\description{
List all GitHub Pages vignettes for one or more GitHub repository.
}
\examples{
vdt <- github_pages_vignettes(owner = "RajLabMSSM", 
                              repo = c("echolocatoR","echodata"))
}
