% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_pages_vignettes_toc.R
\name{github_pages_vignettes_toc}
\alias{github_pages_vignettes_toc}
\title{GitHub Pages: Table of Contents}
\usage{
github_pages_vignettes_toc(vdt, bullet1 = "🦇")
}
\value{
HTML character string.
}
\description{
List all GitHub Pages vignettes for one or more GitHub repository.
Then construct an HTML Table of Contents (TOC) from the vignette links.
}
\keyword{internal}
