% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_repositories.R
\name{github_repositories}
\alias{github_repositories}
\title{GitHub repositories}
\source{
\href{https://docs.github.com/en/rest/search#search-repositories}{
GitHub Docs: Search repositories}

\href{https://docs.github.com/en/rest/search#constructing-a-search-query}{
GitHub Docs: Constructing a search query}

\href{https://docs.github.com/en/search-github/searching-on-github/searching-for-repositories}{
Searching GitHub repos}

\href{https://github.com/orgs/community/discussions/9759}{
Case-sensitive GitHub searches}

\href{https://github.com/r-lib/gh/pull/136}{
Examples of using gh to search repositories}
}
\usage{
github_repositories(
  query,
  token = gh::gh_token(),
  .limit = Inf,
  verbose = TRUE
)
}
\arguments{
\item{query}{Regex query to filter files by.}

\item{token}{GitHub Personal Authentication Token (PAT).
See \href{https://docs.github.com/en/authentication/keeping-your-account-and-data-secure/creating-a-personal-access-token}{
here for further instructions}.}

\item{.limit}{Number of records to return. This can be used
instead of manual pagination. By default it is \code{NULL},
which means that the defaults of the GitHub API are used.
You can set it to a number to request more (or less)
records, and also to \code{Inf} to request all records.
Note, that if you request many records, then multiple GitHub
API calls are used to get them, and this can take a potentially
long time.}

\item{verbose}{Print messages.}
}
\description{
Search for GitHub repositories using specific queries.
}
\examples{
repos <- github_repositories(query="language:r", .limit=100)
}
