% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_workflows.R
\name{github_workflows}
\alias{github_workflows}
\title{GitHub workflows}
\usage{
github_workflows(
  owner,
  repo,
  token = gh::gh_token(),
  latest_only = TRUE,
  workflows = NULL,
  verbose = TRUE
)
}
\arguments{
\item{owner}{Repo owner name
(can be organization or individual GitHub account name).}

\item{repo}{GitHub repository name.}

\item{token}{GitHub Personal Authentication Token (PAT).
See \href{https://docs.github.com/en/authentication/keeping-your-account-and-data-secure/creating-a-personal-access-token}{
here for further instructions}.}

\item{latest_only}{Only return the latest run of each workflow.}

\item{workflows}{Select which workflows to return metadata for.}

\item{verbose}{Print messages.}
}
\value{
\link[data.table]{data.table} containing workflow metadata.
}
\description{
Get metadata on workflows that have been run via GitHub Actions.
This includes the "conclusion" columns showing whether 
the workflow is currently passing.
}
\examples{
dt <- github_workflows(owner="neurogenomics", repo="orthogene")
}
