% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_repos.R
\name{r_repos}
\alias{r_repos}
\title{R repositories}
\usage{
r_repos(save_path = tempfile(pattern = "upsetr.pdf"), verbose = TRUE)
}
\arguments{
\item{save_path}{Path to save upset plot to.}

\item{verbose}{Print messages.}
}
\value{
Named list.
}
\description{
Report on which repositories R packages are distributed through
 (e.g. CRAN, Bioc, rOpenSci, and/or GitHub).
}
\examples{
report <- r_repos()
}
