% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_repos_opts.R
\name{r_repos_opts}
\alias{r_repos_opts}
\title{R repository options}
\usage{
r_repos_opts()
}
\value{
Character vector
}
\description{
A character vector of R packages repositories.
}
\examples{
which <- r_repos_opts()
}
