% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_repos_upset_queries.R
\name{r_repos_upset_queries}
\alias{r_repos_upset_queries}
\title{R repos upset plot queries}
\usage{
r_repos_upset_queries(
  which,
  upset_plot = TRUE,
  highlight = list("GitHub"),
  color = "darkred"
)
}
\arguments{
\item{which}{Which R repositories to extract data from.}

\item{upset_plot}{Whether to create an upset plot 
showing R package overlap between repositories.}

\item{highlight}{Groups to highlight.}

\item{color}{Highlight color.}
}
\value{
Named list of query parameters
}
\description{
Function to generate queries passed to .
}
\examples{
queries <- r_repos_upset_queries(which=NULL)
}
