% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readme_header.R
\name{readme_header}
\alias{readme_header}
\title{Create a README header}
\usage{
readme_header(
  owner = NULL,
  repo = NULL,
  token = gh::gh_token(),
  desc_file = NULL,
  verbose = TRUE,
  branch = github_branches(owner = owner, repo = repo, branch = "master", token = token,
    desc_file = desc_file, verbose = FALSE),
  add_hex = TRUE,
  add_version = TRUE,
  add_actions = TRUE,
  add_commit = TRUE,
  add_codecov = TRUE,
  add_license = TRUE,
  add_authors = TRUE,
  hex_path = "inst/hex/hex.png",
  action = "rworkflows",
  height = 300,
  as_list = FALSE,
  sep = "\\n"
)
}
\arguments{
\item{owner}{Owner of the GitHub repository. 
If \code{NULL}, will automatically try to infer the owner
 name from the \emph{DESCRIPTION file}
 (assuming you're working directory is a local R package repo).}

\item{repo}{GitHub repository name. 
If \code{NULL}, will automatically try to infer the repo name
 name from the \emph{DESCRIPTION file}
 (assuming you're working directory is a local R package repo).}

\item{token}{GitHub Personal Authentication Token (PAT).
See \href{https://docs.github.com/en/authentication/keeping-your-account-and-data-secure/creating-a-personal-access-token}{
here for further instructions}.}

\item{desc_file}{When \code{owner} or \code{repo} are NULL, 
these arguments are inferred from the \emph{DESCRIPTION} file.
If a  \emph{DESCRIPTION} file cannot be found in the current working directory, 
it then searches for one in the remote GitHub repository.}

\item{verbose}{Print messages.}

\item{branch}{[Optional] If \code{branch} is supplied 
(as a character vector of one or more branch names), 
will check to see if that branch exists. If it does, only that branch will
be returned. If it doesn't, an error will be thrown.}

\item{add_hex}{Add a hex sticker with 
\link[echodata]{github_hex}.}

\item{add_version}{Add package version with 
\link[badger]{badge_github_version}.}

\item{add_actions}{Add GitHub Actions status with
\link[badger]{badge_github_actions}.}

\item{add_commit}{Add the last GitHub repo commit date with
\link[badger]{badge_last_commit}.}

\item{add_codecov}{Add CodeCov status with
\link[badger]{badge_codecov}.}

\item{add_license}{Add license info with
\link[badger]{badge_license}.}

\item{add_authors}{Add author names inferred from 
the \code{DESCRIPTION} file.}

\item{hex_path}{Path to the hex sticker in the GitHub repo.}

\item{action}{The name of the GitHub actions workflow. Defaults to
\code{"R-CMD-CHECK"}.}

\item{height}{Height of the HTML image (only used when \code{add_html=TRUE}).}

\item{as_list}{Return the header as a named list (\code{TRUE}), 
or a collapsed text string (default: \code{FALSE}).}

\item{sep}{Character to separate each item in the list with 
using \link[base]{paste}.}
}
\description{
Create a header for your R package's README file,
containing hex stickers and \pkg{badger} badges.
}
\examples{
\dontrun{
    h <- readme_header(owner = "RajLabMSSM",
                       repo = "echolocatoR")
}
}
