% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlist_dt.R
\name{unlist_dt}
\alias{unlist_dt}
\title{Unlist a \link[data.table]{data.table}}
\usage{
unlist_dt(dt, exclude = NULL, verbose = TRUE)
}
\arguments{
\item{dt}{A \link[data.table]{data.table}.}

\item{exclude}{Columns to exclude from unlisting.}

\item{verbose}{Print messages.}
}
\value{
\code{dt} with list columns turned into vectors.
}
\description{
Identify columns that are lists and turn them into vectors.
}
\keyword{internal}
