% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description_extract.R
\name{description_extract}
\alias{description_extract}
\title{DESCRIPTION file: extract fields}
\usage{
description_extract(
  repo = NULL,
  desc_file = NULL,
  fields = c("owner", "repo", "authors"),
  add_html = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{repo}{GitHub repository name.}

\item{desc_file}{When \code{owner} or \code{repo} are NULL, 
these arguments are inferred from the  \emph{DESCRIPTION} file.
If a  \emph{DESCRIPTION} file cannot be found in the current working directory, 
it then searches for one in the remote GitHub repository.}

\item{fields}{Fields to extract.}

\item{add_html}{Add HTML styling to certain fields (e.g "authors").}

\item{verbose}{Print messages.}
}
\description{
Extract information from an R package's DESCRIPTION file.
}
\examples{
 
res <- description_extract(repo="echolocatoR")
}
