% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_commits.R
\name{github_commits}
\alias{github_commits}
\title{GitHub commits}
\usage{
github_commits(owner, repo, token = gh::gh_token(), verbose = TRUE)
}
\arguments{
\item{owner}{Repo owner name
(can be organization or individual GitHub account name).}

\item{repo}{GitHub repository name.}

\item{token}{GitHub Personal Authentication Token (PAT).
See \href{https://docs.github.com/en/authentication/keeping-your-account-and-data-secure/creating-a-personal-access-token}{
here for further instructions}.}

\item{verbose}{Print messages.}
}
\value{
A nested list of metadata
}
\description{
Get metadata for all commits to a given GitHub repo.
}
\examples{
commits <- github_commits(owner="RajLabMSSM", 
                          repo="echolocatoR")
}
