% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_metadata.R
\name{github_metadata}
\alias{github_metadata}
\title{GitHub metadata}
\usage{
github_metadata(
  owner,
  repo,
  add_traffic = FALSE,
  token = gh::gh_token(),
  verbose = TRUE
)
}
\arguments{
\item{owner}{Repo owner name
(can be organization or individual GitHub account name).}

\item{repo}{GitHub repository name.}

\item{add_traffic}{Add traffic metadata 
with \link[echogithub]{github_traffic}.}

\item{token}{GitHub Personal Authentication Token (PAT).
See \href{https://docs.github.com/en/authentication/keeping-your-account-and-data-secure/creating-a-personal-access-token}{
here for further instructions}.}

\item{verbose}{Print messages.}
}
\value{
A \link[data.table]{data.table}.
}
\description{
Get metadata on a GitHub repository.
}
\examples{
dt <- github_metadata(owner="RajLabMSSM", 
                      repo="echolocatoR")
}
