% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_pages.R
\name{github_pages}
\alias{github_pages}
\title{GitHub Pages website}
\usage{
github_pages(
  owner,
  repo,
  error = TRUE,
  token = gh::gh_token(),
  .limit = Inf,
  verbose = TRUE
)
}
\arguments{
\item{owner}{Repo owner name
(can be organization or individual GitHub account name).}

\item{repo}{GitHub repository name.}

\item{error}{Throw an error when the GitHub repository does not exist 
(default: \code{TRUE}).}

\item{token}{GitHub Personal Authentication Token (PAT).
See \href{https://docs.github.com/en/authentication/keeping-your-account-and-data-secure/creating-a-personal-access-token}{
here for further instructions}.}

\item{.limit}{Number of records to return. This can be used
instead of manual pagination. By default it is \code{NULL},
which means that the defaults of the GitHub API are used.
You can set it to a number to request more (or less)
records, and also to \code{Inf} to request all records.
Note, that if you request many records, then multiple GitHub
API calls are used to get them, and this can take a potentially
long time.}

\item{verbose}{Print messages.}
}
\description{
Return the \href{https://pages.github.com/}{GitHub Pages} website URL 
for a given repository, if one exists.
}
\examples{
link <- github_pages(owner="RajLabMSSM",
                     repo="echolocatoR")
}
