% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_traffic.R
\name{github_traffic}
\alias{github_traffic}
\title{GitHub traffic}
\usage{
github_traffic(
  owner,
  repo,
  token = gh::gh_token(),
  na_fill = NULL,
  verbose = TRUE
)
}
\arguments{
\item{owner}{Repo owner name
(can be organization or individual GitHub account name).}

\item{repo}{GitHub repository name.}

\item{token}{GitHub Personal Authentication Token (PAT).
See \href{https://docs.github.com/en/authentication/keeping-your-account-and-data-secure/creating-a-personal-access-token}{
here for further instructions}.}

\item{na_fill}{Value to fill NAs with.}

\item{verbose}{Print messages.}
}
\value{
A \link[data.table]{data.table}.
}
\description{
Get traffic info on a target GitHub repository.
}
\examples{
dt <- github_traffic(owner="RajLabMSSM", 
                     repo="echolocatoR")
}
