% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_repos.R
\name{r_repos}
\alias{r_repos}
\title{R repositories}
\usage{
r_repos(
  which = r_repos_opts(),
  version = NULL,
  add_downloads = FALSE,
  add_descriptions = FALSE,
  add_github = FALSE,
  upset_plot = TRUE,
  show_plot = TRUE,
  queries = r_repos_upset_queries(which = which, upset_plot = upset_plot),
  save_path = tempfile(fileext = "upsetr.pdf"),
  height = 7,
  width = 10,
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{which}{Which R repositories to extract data from.}

\item{version}{(Optional) \code{character(1)} or \code{package_version}
indicating the \emph{Bioconductor} version (e.g., "3.8") for which
repositories are required.}

\item{add_downloads}{Add the number of downloads from each repository.}

\item{add_descriptions}{Add metadata from \emph{DESCRIPTION} files.}

\item{add_github}{Add metadata from the respective GitHub repository 
for each R package (if any exists).}

\item{upset_plot}{Whether to create an upset plot 
showing R package overlap between repositories.}

\item{show_plot}{Print the plot.}

\item{queries}{Unified query of intersections, elements, and custom row functions. Entered as a list that contains a list of
queries. query is the type of query being conducted. params are the parameters of the query (if any). color is the color of the points on the
plot that will represent the query. If no color is selected one will be provided automatically. active takes TRUE or FALSE, and if
TRUE, it will overlay the bars present  with the results from the query. If FALSE a tick mark will indicate the intersection size.
See examples section on how to do this.}

\item{save_path}{Path to save upset plot to.}

\item{height}{Saved plot height.}

\item{width}{Saved plot width.}

\item{nThread}{Number of threads to parallelise data queries across.}

\item{verbose}{Print messages.}
}
\value{
Named list.
}
\description{
Report on which repositories R packages are distributed through
 (i.e. base R, CRAN, Bioc, rOpenSci, R-Forge, and/or GitHub).
}
\examples{
report <- r_repos()
}
