% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_repos_data.R
\name{r_repos_data}
\alias{r_repos_data}
\title{R repositories data}
\usage{
r_repos_data(
  add_downloads = FALSE,
  which = r_repos_opts(),
  version = NULL,
  verbose = TRUE
)
}
\arguments{
\item{add_downloads}{Add the number of downloads from each repository.}

\item{which}{Which R repositories to extract data from.}

\item{version}{(Optional) \code{character(1)} or \code{package_version}
indicating the \emph{Bioconductor} version (e.g., "3.8") for which
repositories are required.}

\item{verbose}{Print messages.}
}
\value{
data.table
}
\description{
Gather data on which repositories R packages are distributed through
 (e.g. CRAN, Bioc, rOpenSci, and/or GitHub).
}
\examples{
pkgs <- r_repos_data()
}
