% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_repos_data.R
\name{r_repos_data}
\alias{r_repos_data}
\title{R repositories data}
\usage{
r_repos_data(
  include = NULL,
  add_downloads = FALSE,
  add_descriptions = FALSE,
  add_github = FALSE,
  add_hex = FALSE,
  hex_path = "inst/hex/hex.png",
  which = r_repos_opts(),
  cast = FALSE,
  fields = NULL,
  version = NULL,
  nThread = 1,
  token = gh::gh_token(),
  verbose = TRUE
)
}
\arguments{
\item{include}{A subset of packages to return data for.}

\item{add_downloads}{Add the number of downloads from each repository.}

\item{add_descriptions}{Add metadata from \emph{DESCRIPTION} files.}

\item{add_github}{Add metadata from the respective GitHub repository 
for each R package (if any exists).}

\item{add_hex}{Search for hex sticker URLs for each package.
If the link does not exist, returns \code{NA}.}

\item{hex_path}{Path to hex sticker file.}

\item{which}{Which R repositories to extract data from.}

\item{cast}{Cast the results to wide format
so that each package only appears in one row.}

\item{fields}{Fields to extract.}

\item{version}{(Optional) \code{character(1)} or \code{package_version}
indicating the \emph{Bioconductor} version (e.g., "3.8") for which
repositories are required.}

\item{nThread}{Number of threads to parallelise \code{add_descriptions} 
step across.}

\item{token}{GitHub Personal Authentication Token (PAT).
See \href{https://docs.github.com/en/authentication/keeping-your-account-and-data-secure/creating-a-personal-access-token}{
here for further instructions}.}

\item{verbose}{Print messages.}
}
\value{
data.table
}
\description{
Gather data on which repositories R packages are distributed through
 (e.g. CRAN, Bioc, rOpenSci, and/or GitHub).
}
\examples{
#### All packages ####
pkgs1 <- r_repos_data()
#### Specific packages ####
#### Add extra data ####
pkgs2 <- r_repos_data(include=c("echogithub","echodeps","BiocManager"),
                      which = r_repos_opts(exclude = NULL),
                      add_downloads = TRUE,
                      add_descriptions = TRUE,
                      add_github = TRUE,
                      add_hex = TRUE,
                      cast = TRUE)
}
