% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_repos_downloads.R
\name{r_repos_downloads}
\alias{r_repos_downloads}
\title{Get downloads data from CRAN/Bioconductor/GitHub}
\usage{
r_repos_downloads(
  pkgs,
  which = r_repos_opts(),
  use_cache = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{pkgs}{Output from \link[echogithub]{r_repos_data}.}

\item{which}{Which R repositories to extract data from.}

\item{use_cache}{logical, should cached data be used? Default: TRUE. If set to FALSE, it will
re-query download stats and update cache.}

\item{verbose}{Print messages.}
}
\description{
Get the total number of downloads for each R package in 
CRAN, Bioconductor, and/or GitHub.
}
\examples{
pkgs <- r_repos_data()[r_repo=="CRAN",][seq_len(5),]
pkgs2 <- r_repos_downloads(pkgs = pkgs)
}
